requirements.txt
autoenv==1.0.0
backports.csv==1.0.5
cli-helpers==0.2.3
click==6.7
configobj==5.0.6
Flask==0.12.2
gunicorn==19.7.1
humanize==0.5.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pgcli==1.8.1
pgspecial==1.9.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
Pygments==2.2.0
python-dotenv==0.7.1
setproctitle==1.1.10
six==1.11.0
SQLAlchemy==1.1.15
sqlparse==0.2.4
terminaltables==3.1.0
uWSGI==2.0.15
wcwidth==0.1.7
Werkzeug==0.12.2
